requirements.txt
absl-py==2.1.0
asttokens==2.2.1
attrs==24.1.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
flatbuffers==24.3.25
fonttools==4.39.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.2
ipython==8.10.0
jax==0.4.13
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.6
numpy==1.23.4
opencv-contrib-python==4.10.0.84
opencv-python==4.7.0.68
openpyxl==3.1.1
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pandas-datareader==0.10.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.0.0
prompt-toolkit==3.0.37
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.22
pygame==1.9.6
Pygments==2.14.0
PyMuPDF==1.23.15
PyMuPDFb==1.23.9
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305; sys_platform == "win32"
pyzmq==25.0.0
requests==2.28.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sounddevice==0.4.7
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.6
zipp==3.15.0
websockets==12.0
google-cloud-speech==2.24.0
openai==1.54.4
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
pymysql==1.1.0
pyjwt==2.8.0
werkzeug==3.0.0
mysql-connector-python==8.2.0
pydub==0.25.1
pdf2image==1.16.3
